LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-readonly-subfield-expected.txt

Sub-fields in input[type=time] should be read-only in some cases. This requires window.internals.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


createTimeInput argument order: min, step, value

Milliseconds field:
PASS isReadOnlyField(createTimeInput("00:00:00.500", step1sec, ""), pseudoMillisecond) is true
PASS isReadOnlyField(createTimeInput("00:00:00.500", step1sec, "00:00:00.500"), pseudoMillisecond) is true
PASS isReadOnlyField(createTimeInput("00:00:00.500", step1sec, "00:00:00.600"), pseudoMillisecond) is false
PASS isReadOnlyField(createTimeInput("00:00:00.000", step1sec, "00:00:00.600"), pseudoMillisecond) is false
Seconds field:
PASS isReadOnlyField(createTimeInput("00:00:30", step1min, ""), pseudoSecond) is true
PASS isReadOnlyField(createTimeInput("00:00:30", step1min, "00:00:30"), pseudoSecond) is true
PASS isReadOnlyField(createTimeInput("00:00:30.500", step1min, "00:00:30.600"), pseudoSecond) is true
PASS isReadOnlyField(createTimeInput("00:00:30", step1min, "00:00:35"), pseudoSecond) is false
PASS isReadOnlyField(createTimeInput("00:00:00", step1min, "00:00:35"), pseudoSecond) is false
Minutes field:
PASS isReadOnlyField(createTimeInput("00:30", step1hour, ""), pseudoMinute) is true
PASS isReadOnlyField(createTimeInput("00:30", step1hour, "00:30"), pseudoMinute) is true
PASS isReadOnlyField(createTimeInput("00:30", step1hour, "00:35"), pseudoMinute) is false
PASS isReadOnlyField(createTimeInput("00:00", step1hour, "00:35"), pseudoMinute) is false
Hour field:
PASS isReadOnlyField(createTimeInput("11:00", step1day, ""), pseudoHour) is false
PASS isReadOnlyField(createTimeInput("11:00", step1day, ""), pseudoAMPM) is false
PASS isReadOnlyField(createTimeInput("11:00", step1day, ""), pseudoMinute) is true
PASS isReadOnlyField(createTimeInput("11:30", step1day, "11:30"), pseudoHour) is false
PASS isReadOnlyField(createTimeInput("11:30", step1day, "11:30"), pseudoAMPM) is false
PASS isReadOnlyField(createTimeInput("11:30", step1day, "11:30"), pseudoMinute) is true
PASS isReadOnlyField(createTimeInput("11:30", step1day, "12:30"), pseudoHour) is false
PASS isReadOnlyField(createTimeInput("11:30", step1day, "12:30"), pseudoAMPM) is false
PASS isReadOnlyField(createTimeInput("11:30", step1day, "12:30"), pseudoMinute) is true
PASS isReadOnlyField(createTimeInput("00:00", step1day, "12:30"), pseudoHour) is false
PASS isReadOnlyField(createTimeInput("00:00", step1day, "12:30"), pseudoAMPM) is false
PASS isReadOnlyField(createTimeInput("00:00", step1day, "12:30"), pseudoMinute) is false
PASS successfullyParsed is true

TEST COMPLETE
